ci, scale: Add EGW masquerade delay scale test
This commit introduces a new scale test for Egress Gateway. At a high
level, the scale test creates a series of Pods that attempt to
communicate to an external target through an EGW node. Metrics are
tracked by each client to record the client's experience in connecting
to the external target, such as how long it took for a connection to be
masqueraded through an EGW node, and how many connections could be
established before masquerading took place.

An extended description of the benchmark can be found in the
[cilium/scaffolding][1] repository, which hosts the benchmark's tooling.

The GHA workflow in this commit creates an EKS cluster to host the test.
A node is deployed without Cilium to host the external target. Nodes
hosting client Pods are configured to have a pod limit of 100, which
respects the upstream Kubernetes guidelines of 110 pods per node while
adding some buffer. Cilium is installed with prefix delegation enabled
to help limit the number of nodes that are required to be deployed. The
workflow will run the masquerade delay test, as well as the masquerade
delay baseline test. The latter is the exact same test as the former,
but with no EGW policy deployed, causing client pods to connect directly
to the external target without being routed through an EGW node.

[1]: https://github.com/cilium/scaffolding/blob/main/egw-scale-utils/README.md

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>